Fix #393: Fix dimension check for quarter isocentre leaves
Previously, the code exited with an error message if ZTONGUE was
> ZGROOVE for quarter isocentre leaves.  Confusingly, the output message
stated that the code was exiting because ZTONGUE<ZGROOVE. Ultimately, it
was determined that the output message was correct and the check should
be on ZTONGUE<ZGROOVE.  Bug found by Marc Chamberland.  Fixed with info
provided by Emily Heath and Tony Popescu.